Create LINKS (work in progress)
This list is a list of various search engines and links I have gathered over the years. As a hobby, collecting search engines and ways to travel the internet is quite interesting. Google and other companies have too much of a monopoly over where we get our information and what it is. This is what started me off.  I have quite a lot of other links too add and until now have simply bookmarked them en browser or emailed them to myself, but this is impractical and quite selfish without any proper way of getting at them or sharing them. 
So I hope by putting it HERE it will have some use for other people and maybe I could get some help with putting this project together and weeding out the bad links, the Google dependent engines and the quite useless websites.

Suggestions and additions are VERY welcome and to those who have used Github longer than me; a little help would be much appreciated :-) 
Enjoy and Thank You in advance!!